requirements.txt
pillow
playwright
requests
